Changes:
- Rename some endpoints and props. Mostly the
ones used by Grades and Schedule components.
- Add Schedule comonent, ath this momment only
for professor. It's completede
- Add Evaluation Functionality for professor.
At this momment is pendng the group notes and
notes creation, and note management by student.
- Adding more utility functions
- The fetching process in the Schedule and
evaluation comonents its exactly the same, it
was Copy/paste from one onto another. This will
need to be fixed in a near future.
- More changes that i forgot, but the main ones
were the above described